pyproject.toml
[tool.poetry]
name = "tensor-manager"
version = "0.1.0"
description = "unshittify nvidia GPU virtual memory protection bullshit"
authors = ["Jacob Howe <howejacob1@gmail.com>"]
readme = "README.md"
packages = [{include = "tensor_manager"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
torch = {version = "2.9.0+cu128", source = "torch-cu128"}
torchvision = {version = "0.24.0+cu128", source = "torch-cu128"}
torchaudio = {version = "2.9.0+cu128", source = "torch-cu128"}
cupy-cuda12x = "^13.6.0"
cuda-python = "^12.6.0"
fastapi = "^0.121.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
python-multipart = "^0.0.20"
pytest = "^8.4.2"
httpx = "^0.28.1"
numpy = "^2.3.4"
requests = "^2.32.3"
soundfile = "^0.13.1"


[[tool.poetry.source]]
name = "torch-cu128"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
